+++ R/input-checkboxgroup.R
@@ -94,10 +94,14 @@ checkboxGroupInput <- function(inputId, label, choices = NULL, selected = NULL,
     divClass <- paste(divClass, "shiny-input-container-inline")
 
   # return label and select tag
+  inputLabel <- shinyInputLabel(inputId, label)
   tags$div(id = inputId,
     style = css(width = validateCssUnit(width)),
     class = divClass,
-    shinyInputLabel(inputId, label),
+    # https://www.w3.org/TR/wai-aria-practices/examples/checkbox/checkbox-1/checkbox-1.html
+    role = "group",
+    `aria-labelledby` = inputLabel$attribs$id,
+    inputLabel,
     options
   )
 }