+++ tests/testthat/test-bootstrap.r
@@ -323,7 +323,7 @@ test_that("normalizeChoicesArgs does its job", {
 
 test_that("Choices need not be provided, can be NULL or c()", {
 
-  expected <- "<div id=\"cb\" class=\"form-group shiny-input-checkboxgroup shiny-input-container\">\n  <label class=\"control-label\" id=\"cb-label\" for=\"cb\">Choose:</label>\n  <div class=\"shiny-options-group\"></div>\n</div>"
+  expected <- '<div id="cb" class="form-group shiny-input-checkboxgroup shiny-input-container" role="group" aria-labelledby="cb-label">\n  <label class="control-label" id="cb-label" for="cb">Choose:</label>\n  <div class="shiny-options-group"></div>\n</div>'
   noChoices <- checkboxGroupInput("cb", "Choose:")
   choicesNull <- checkboxGroupInput("cb", "Choose:", choices = NULL)
   choicesCharacter <- checkboxGroupInput("cb", "Choose:", choices = c())
@@ -337,4 +337,25 @@ test_that("Choices need not be provided, can be NULL or c()", {
   expect_identical(noChoices, allChoicesNull)
 
   expect_equal(expected, format(noChoices))
+})
+
+# https://github.com/rstudio/shiny/pull/3187
+test_that("radioButtons() and checkboxGroupInput() accessibility", {
+  rb <- radioButtons("foo", "bar", c("a", "b"))
+  rb_lbl <- rb$children[[1]]
+  expect_equal(rb$attribs$role, "radiogroup")
+  expect_equal(rb_lbl$name, "label")
+  expect_true(!is.null(rb_lbl$attribs$id))
+  expect_equal(
+    rb$attribs$`aria-labelledby`, rb_lbl$attribs$id
+  )
+
+  cbg <- checkboxGroupInput("foo", "bar", c("a", "b"))
+  cbg_lbl <- cbg$children[[1]]
+  expect_equal(cbg$attribs$role, "group")
+  expect_equal(cbg_lbl$name, "label")
+  expect_true(!is.null(cbg_lbl$attribs$id))
+  expect_equal(
+    cbg$attribs$`aria-labelledby`, cbg_lbl$attribs$id
+  )
 })